Revise conn err handling to prevent possible deadlocks
errors channel now initialized before passing it to FSConn. It now re-
presents the only means of communication between the readEvents gorou-
tine and the goroutine handling errors coming from readEvents. That's
because manual shutdowns from sources external to fsock library can be
detected by checking the type of the error (shutting down manually
means closing the connection, which triggers a net.OpError to be
returned).

For the above reason, stopReadEvents and readEventsFSStopped channels
have been dropped for being redundant.

The logic for the goroutine that's handling readEvents errors has been
moved to its own function.

We are now making sure to also send nil errors in case of graceful
shutdowns and successful reconnects on the stopError channel.

Added descriptive comments where necessary.