Handle ECONNRESET error to trigger reconnect in readHeaders
Previously, all OpErrors were considered intentional and signaled a
graceful shutdown, ignoring 'connection reset by peer' errors wrapped
within OpErrors. This prevented attempts to reconnect to FreeSWITCH.

syscall.ECONNRESET is now also changed to io.EOF before sending it over
to the conn err handler, making sure that a reconnect is triggered.

A test was added to make sure we account for this issue in the future.